refactor: extract SourceCard's badge/date formatting to a shared module
DecisionTimeline is about to become a second consumer of badgeText and
relativeDate; per DRY, pull them into lib/sourceFormat.js instead of
reimplementing or importing internals across modules.
